Don't use PB3 for the SPI bus, because that pin is used for ITM!
ITM broke when PB3 was used for SPI, this is because the default
configuration for PB3 is SWO (used for ITM tracing), and configuring it
as an SPI clock pin instead (even tho the WS28xx doesn't actually use
that signal) disconnected the internal ITM / SWO peripheral from PB3,
preventing its use.

Because this is a boot-time configuration, the Rust type system doesn't
know about this (possibly the ITM could be accessed via stm32f4xx_hal
instead, and stop this mistake being made (at compile time)).